finishing all CRUD functionality for posts

+++ app/controllers/Posts.php
@@ -134,6 +134,14 @@
 
     public function delete($id){
       if($_SERVER['REQUEST_METHOD'] == 'POST'){
+        $post = $this->postModel->getPostById($id);
+
+        // Check for owner
+        if($post->user_id != $_SESSION['user_id']){
+          flash('You may only delete your own posts');
+          redirect('posts');
+        }
+
         if($this->postModel->deletePost($id)){
           flash('post_message', 'Post Removed');
           redirect('posts');
